backend/requirements.txt
pyht
pandas
python-dotenv
google-generativeai